requirements.txt
# Production dependencies (required for Firebase Functions)
flask==2.3.3
gunicorn==21.2.0
firebase-admin==6.2.0
twilio==8.6.0
sendgrid==6.10.0
python-dotenv==1.0.0
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==4.0.4
requests==2.31.0
functions-framework==3.4.0
Werkzeug==2.3.7

# Development dependencies (not deployed to Firebase Functions)
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
isort==5.12.0